move all global variables to global_vars file
